+++ scripts/utils.js
@@ -1,0 +1,14 @@
+module.exports = {
+    /**
+     * Converts a brand title into a filename (not a full path)
+     * @param {String} title The title to convert
+     */
+    titleToFilename: title => (
+        title.toLowerCase()
+            .replace(/\+/g, "plus")
+            .replace(/^\./, "dot-")
+            .replace(/\.$/, "-dot")
+            .replace(/\./g, "-dot-")
+            .replace(/[ !’]/g, '')
+    )
+}
